Partial TS rewrite
