+++ ts/flexbuffers/value-type.ts
@@ -1,0 +1,30 @@
+export enum ValueType {
+    NULL = 0,
+    INT = 1,
+    UINT = 2,
+    FLOAT = 3,
+    KEY = 4,
+    STRING = 5,
+    INDIRECT_INT = 6,
+    INDIRECT_UINT = 7,
+    INDIRECT_FLOAT = 8,
+    MAP = 9,
+    VECTOR = 10,
+    VECTOR_INT = 11,
+    VECTOR_UINT = 12,
+    VECTOR_FLOAT = 13,
+    VECTOR_KEY = 14,
+    VECTOR_STRING_DEPRECATED = 15,
+    VECTOR_INT2 = 16,
+    VECTOR_UINT2 = 17,
+    VECTOR_FLOAT2 = 18,
+    VECTOR_INT3 = 19,
+    VECTOR_UINT3 = 20,
+    VECTOR_FLOAT3 = 21,
+    VECTOR_INT4 = 22,
+    VECTOR_UINT4 = 23,
+    VECTOR_FLOAT4 = 24,
+    BLOB = 25,
+    BOOL = 26,
+    VECTOR_BOOL = 36,
+}
